Change parser and command to fit Clikt library
The Parser creates Commands, which return CommandOutputs upon execution.
This functionality is not supported by Clikt, in which both parsing and execution are
done by the CliktCommand object, and commands can only return exit status
codes. Let's create a Command class in which Clikt objects are created and
executed. This allows most existing functionality to be retained, while also
allowing for the use of the Clikt library
CliktCommand instance